Enabled sending of the PropertiesChanged Signal.
A class is eligible to send a PropertiesChanged signal if its interface
implements the INotifyPropertyChanged interface. Only then will the
code-generator create the necessary method to handle the events.
Any property for which the signal will be emitted has to be marked with
the DbusPropertiesChanged attribute, else the property will not be
recognized by the eventhandler method.